refactor(shell): share command parsing and inspection metadata
