+++ packages/shell/src/workspace/command-analysis.ts
@@ -1,3 +1,5 @@
+import { parseShellCommand, splitShellCommandSegments } from "../command/parse.ts"
+
 export interface WorkspaceInspectionCommandSegment {
   command: string
   followsPipe: boolean
@@ -9,7 +11,7 @@ export interface WorkspaceInspectionCommandSegment {
 
 export function analyzeWorkspaceInspectionCommand(command: string): WorkspaceInspectionCommandSegment[] {
   return splitShellCommandSegments(command).map((segment) => {
-    const words = parseShellWords(segment.command)
+    const words = parseShellCommand(segment.command, "inspection")
     return {
       ...segment,
       paths: shellPathArguments(words),
@@ -226,61 +228,6 @@ function pathArgumentsUntilShellBoundary(args: string[]) {
   return paths
 }
 
-function splitShellCommandSegments(command: string) {
-  const segments: Array<{ command: string, followsPipe: boolean, separatorAfter?: "&&" | "||" | "|" | ";" | "\n" }> = []
-  let current = ""
-  let quote: "'" | "\"" | undefined
-  let escaped = false
-  let followsPipe = false
-  for (let index = 0; index < command.length; index++) {
-    const char = command[index]!
-    if (escaped) {
-      current += char
-      escaped = false
-      continue
-    }
-    if (char === "\\") {
-      current += char
-      escaped = true
-      continue
-    }
-    if (quote) {
-      if (char === quote) quote = undefined
-      current += char
-      continue
-    }
-    if (char === "'" || char === "\"") {
-      quote = char
-      current += char
-      continue
-    }
-    const next = command[index + 1]
-    if (char === "&" && next === "&") {
-      segments.push({ command: current, followsPipe, separatorAfter: "&&" })
-      current = ""
-      followsPipe = false
-      index += 1
-      continue
-    }
-    if (char === "|" && next === "|") {
-      segments.push({ command: current, followsPipe, separatorAfter: "||" })
-      current = ""
-      followsPipe = false
-      index += 1
-      continue
-    }
-    if (char === "|" || char === ";" || char === "\n") {
-      segments.push({ command: current, followsPipe, separatorAfter: char })
-      current = ""
-      followsPipe = char === "|"
-      continue
-    }
-    current += char
-  }
-  segments.push({ command: current, followsPipe })
-  return segments
-}
-
 function hasRecursiveGrepFlag(words: string[]) {
   const args = words.slice(1)
   for (let index = 0; index < args.length; index++) {
@@ -299,41 +246,4 @@ function hasRecursiveGrepFlag(words: string[]) {
     if (arg.includes("r") || arg.includes("R")) return true
   }
   return false
-}
-
-function parseShellWords(command: string) {
-  const words: string[] = []
-  let current = ""
-  let quote: "'" | "\"" | undefined
-  let escaped = false
-  for (const char of command) {
-    if (escaped) {
-      current += char
-      escaped = false
-      continue
-    }
-    if (char === "\\") {
-      escaped = true
-      continue
-    }
-    if (quote) {
-      if (char === quote) quote = undefined
-      else current += char
-      continue
-    }
-    if (char === "'" || char === "\"") {
-      quote = char
-      continue
-    }
-    if (/\s/.test(char)) {
-      if (current) {
-        words.push(current)
-        current = ""
-      }
-      continue
-    }
-    current += char
-  }
-  if (current) words.push(current)
-  return words
 }
